refactor: Change unit detection logic to work with monorepo
When the subdirectory lacks a unitindex.md file, look in the
subdirectory for markdown files. Following the naming convention
(u1ws.md, u2ws.md, etc.), the numbers in filenames will be extracted,
sorted, and counted to determine the number of units.

chore: Clean up comments and debug fn and statements