fix: Inconsistent exit code on pip show when checking for deps
The exit code of the `pip show | grep` that I was using to check
if the dependencies were already installed was inconsistent. As a
result, it would sometimes result in skipping the dependency
installation process.

I removed that guard and continued to run `pip install`, as
there's no downside to doing so.

Additionally fixed the `trap` which was double firing in some instances.